CA-172369: Separate syslog package and use "-v" build flag.
